- Added a new header file to hold the augmenta data
